+++ src/useFocusContext.ts
@@ -1,8 +1,7 @@
-import { useContext, createContext } from 'react';
+import { createContext, useContext } from 'solid-js';
 import { ROOT_FOCUS_KEY } from './SpatialNavigation';
 
-export const FocusContext = createContext(ROOT_FOCUS_KEY);
-FocusContext.displayName = 'FocusContext';
+export const FocusContext = createContext<string>(ROOT_FOCUS_KEY);
 
 /** @internal */
-export const useFocusContext = () => useContext(FocusContext);
+export const useFocusContext = (): string => useContext(FocusContext) ?? ROOT_FOCUS_KEY;