+++ webpack.config.prod.js
@@ -16,7 +16,7 @@ module.exports = {
   resolve: {
     extensions: ['.tsx', '.ts', '.js', '.jsx']
   },
-  externals: ['react', /^lodash(\/.+)?$/],
+  externals: ['solid-js', /^lodash(\/.+)?$/],
   output: {
     filename: 'index.js',
     path: path.resolve(__dirname, 'dist'),